Added to dev environment
I was concerned about speed due to the large amount of comments and
debug statements that the script was required to parse through. As a
result, I created “build_clean_bashttpd.sh” to strip out all the
comments and debug logging features, and write that to the main
“bashttpd.sh” file. Meanwhile, “dev_bashttpd.sh” will be maintained
with all it’s comments and debug logging in order to use those features
during development.